requirements.txt
mysqlclient
xlsxwriter
pyserial
